requirements.txt
fastapi~=0.111.0
uvicorn
pydantic~=2.7.4
sqlalchemy~=2.0.31
psycopg2-binary
python-dotenv~=1.0.1